Fix return value of gettimeofday implementation
Return value was changed to int in the header only.
This generated a multiple definitions error, breaking the build process.

+++ source/system_calls/gettimeofday.c
@@ -2,7 +2,7 @@
 #include <liblinux/system_calls/gettimeofday.h>
 #include <liblinux/definitions.h>
 
-long gettimeofday(struct timeval *tv, struct timezone *tz)
+int gettimeofday(struct timeval *tv, struct timezone *tz)
 {
 	return system_call(__NR_gettimeofday, tv, tz);
 }
